+++ Source/ProjectZ/Private/Character/ZCharacter.cpp
@@ -35,43 +35,3 @@ AZCharacter::AZCharacter()
 	FirstPersonCamera->bUsePawnControlRotation = true;
 }
 
-// When PlayerController Possess the Character
-void AZCharacter::PossessedBy(AController* NewController)
-{
-	Super::PossessedBy(NewController);
-	if (bOnPossessActivated) return;
-	bOnPossessActivated = true;
-	InitAbilityActorInfo();
-	AddCharacterAbility();
-}
-
-// Setup Character
-void AZCharacter::InitAbilityActorInfo()
-{
-	if (AbilitySystemComponent) return;
-	AZPlayerState* ZPlayerState = GetPlayerState<AZPlayerState>();
-	check(ZPlayerState);
-	ZPlayerState->GetAbilitySystemComponent()->InitAbilityActorInfo(ZPlayerState, this);
-	Cast<UZAbilitySystemComponent>(ZPlayerState->GetAbilitySystemComponent())->AbilityActorInfoSet();
-	AbilitySystemComponent = ZPlayerState->GetAbilitySystemComponent();
-	AttributeSet = ZPlayerState->GetAttributeSet();
-
-	if (AZNonCombatPlayerController* ZPlayerController = Cast<AZNonCombatPlayerController>(GetController()))
-	{
-		if (AZHUDBase* ZHUD = Cast<AZHUDBase>(ZPlayerController->GetHUD()))
-		{
-			ZHUD->InitOverlay(ZPlayerController, GetPlayerState(), GetAbilitySystemComponent(), GetAttributeSet(), this);
-		}
-	}
-	InitializeDefaultAttributes();
-
-	// TO DO : Make Difference by game mode
-	CardComponent->InitializeNonCombat(this);
-	CardComponent->InitializeCombat(this);
-}
-
-int32 AZCharacter::GetLevel()
-{
-	const AZPlayerState* ZPlayerState = GetPlayerState<AZPlayerState>();
-	return ZPlayerState->GetPlayerLevel();
-}
