+++ Source/ProjectZ/Private/Player/ZNonCombatPlayerController.cpp
@@ -30,13 +30,9 @@ void AZNonCombatPlayerController::SetupInputComponent()
 {
 	Super::SetupInputComponent();
 
-	UZInputComponent* ZInputComponent = CastChecked<UZInputComponent>(InputComponent);
-
 	ZInputComponent->BindAction(MoveAction, ETriggerEvent::Triggered, this, &AZNonCombatPlayerController::Move);
 	ZInputComponent->BindAction(LookAction, ETriggerEvent::Triggered, this, &AZNonCombatPlayerController::Look);
-
-	// Bind Input Action with tags
-	ZInputComponent->BindAbilityAction(InputConfig, this, &AZNonCombatPlayerController::AbilityInputTagPressed, &AZNonCombatPlayerController::AbilityInputTagReleased, &AZNonCombatPlayerController::AbilityInputTagHeld);
+	
 }
 
 void AZNonCombatPlayerController::Move(const FInputActionValue& Value)
@@ -59,30 +55,4 @@ void AZNonCombatPlayerController::Look(const FInputActionValue& InputActionValue
 		ControlledPawn->AddControllerPitchInput(LookAxisVector.Y);
 		ControlledPawn->AddControllerYawInput(LookAxisVector.X);
 	}
-}
-
-void AZNonCombatPlayerController::AbilityInputTagPressed(FGameplayTag InputTag)
-{
-	//GEngine->AddOnScreenDebugMessage(1, 3.f, FColor::Red, *InputTag.ToString());
-}
-
-void AZNonCombatPlayerController::AbilityInputTagReleased(FGameplayTag InputTag)
-{
-	if (GetASC() == nullptr) return;
-	GetASC()->AbilityInputTagReleased(InputTag);
-}
-
-void AZNonCombatPlayerController::AbilityInputTagHeld(FGameplayTag InputTag)
-{
-	if (GetASC() == nullptr) return;
-	GetASC()->AbilityInputTagHeld(InputTag);
-}
-
-UZAbilitySystemComponent* AZNonCombatPlayerController::GetASC()
-{
-	if (ZAbilitySystemComponent == nullptr)
-	{
-		ZAbilitySystemComponent = Cast<UZAbilitySystemComponent>(UAbilitySystemBlueprintLibrary::GetAbilitySystemComponent(GetPawn<APawn>()));
-	}
-	return ZAbilitySystemComponent;
 }
