+++ Source/ProjectZ/Public/Character/ZCharacter.h
@@ -3,28 +3,21 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Character/ZCharacterBase.h"
+#include "ZPlayerCharacter.h"
 #include "ZCharacter.generated.h"
 
 class UCameraComponent;
 class UCardComponent;
 
 UCLASS()
-class PROJECTZ_API AZCharacter : public AZCharacterBase
+class PROJECTZ_API AZCharacter : public AZPlayerCharacter
 {
 	GENERATED_BODY()
 
 public:
 	AZCharacter();
 
-	virtual void PossessedBy(AController* NewController) override;
-	virtual void InitAbilityActorInfo() override;
-
-	virtual int32 GetLevel() override;
-
 private:
 	UPROPERTY(VisibleAnywhere, Category = Camera)
 	UCameraComponent* FirstPersonCamera;
-
-	bool bOnPossessActivated = false;
 };
